Move classes in boosting/cpp/losses_label_wise.h/cpp and boosting/cpp/losses_example_wise.h/cpp to separate header/implementation files.
