+++ python/boomer/boosting/cpp/losses/loss_example_wise_logistic.h
@@ -1,0 +1,24 @@
+/**
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "loss_example_wise.h"
+
+
+namespace boosting {
+
+    /**
+     * A multi-label variant of the logistic loss that is applied example-wise.
+     */
+    class ExampleWiseLogisticLoss : public IExampleWiseLoss {
+
+        public:
+
+            void updateExampleWiseStatistics(uint32 exampleIndex, const IRandomAccessLabelMatrix& labelMatrix,
+                                             const DenseNumericMatrix<float64>& scoreMatrix,
+                                             DenseExampleWiseStatisticMatrix& statisticMatrix) const override;
+
+    };
+
+}
