+++ python/boomer/boosting/cpp/losses/loss_label_wise_squared_error.cpp
@@ -1,0 +1,12 @@
+#include "loss_label_wise_squared_error.h"
+
+using namespace boosting;
+
+
+void LabelWiseSquaredErrorLoss::updateGradientAndHessian(DenseVector<float64>::iterator gradient,
+                                                         DenseVector<float64>::iterator hessian, bool trueLabel,
+                                                         float64 predictedScore) const {
+    float64 expectedScore = trueLabel ? 1 : -1;
+    *gradient = (2 * predictedScore) - (2 * expectedScore);
+    *hessian = 2;
+}
