+++ python/boomer/boosting/cpp/losses/loss_label_wise_squared_error.h
@@ -1,0 +1,24 @@
+/**
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "loss_label_wise_common.h"
+
+
+namespace boosting {
+
+    /**
+     * A multi-label variant of the squared error loss that is applied label-wise.
+     */
+    class LabelWiseSquaredErrorLoss : public AbstractLabelWiseLoss {
+
+        protected:
+
+            void updateGradientAndHessian(DenseVector<float64>::iterator gradient,
+                                          DenseVector<float64>::iterator hessian, bool trueLabel,
+                                          float64 predictedScore) const override;
+
+    };
+
+}
