+++ python/boomer/boosting/cpp/statistics/statistics_label_wise.h
@@ -8,8 +8,8 @@
 #pragma once
 
 #include "../../../common/cpp/statistics/statistics.h"
+#include "../losses/loss_label_wise.h"
 #include "../rule_evaluation_label_wise.h"
-#include "../losses_label_wise.h"
 
 
 namespace boosting {
