+++ python/setup.py
@@ -64,11 +64,13 @@ sources = [
     'boomer/boosting/cpp/data/matrix_dense_example_wise.cpp',
     'boomer/boosting/cpp/data/vector_dense_label_wise.cpp',
     'boomer/boosting/cpp/data/vector_dense_example_wise.cpp',
+    'boomer/boosting/cpp/losses/loss_label_wise_common.cpp',
+    'boomer/boosting/cpp/losses/loss_label_wise_logistic.cpp',
+    'boomer/boosting/cpp/losses/loss_label_wise_squared_error.cpp',
+    'boomer/boosting/cpp/losses/loss_example_wise_logistic.cpp',
     'boomer/boosting/cpp/post_processing/shrinkage_constant.cpp',
     'boomer/boosting/cpp/blas.cpp',
     'boomer/boosting/cpp/lapack.cpp',
-    'boomer/boosting/cpp/losses_label_wise.cpp',
-    'boomer/boosting/cpp/losses_example_wise.cpp',
     'boomer/boosting/cpp/rule_evaluation_label_wise.cpp',
     'boomer/boosting/cpp/rule_evaluation_example_wise.cpp',
     'boomer/boosting/cpp/statistics/statistics_label_wise.cpp',
